Bitmap 끘

+++ Bitmap/Timer.cpp
@@ -45,7 +45,7 @@ bool Timer::Render()
 {
 	SetBkColor(g_hOffScreenDC, RGB(255, 255, 153));
 	SetTextColor(g_hOffScreenDC, RGB(255, 0, 255));
-	TextOut(g_hOffScreenDC, 0, 0, m_csBuffer, _tcslen(m_csBuffer));
+	TextOut(g_hOffScreenDC, 0, 0, m_csBuffer, (int)_tcslen(m_csBuffer));
 	return true;
 }
 bool Timer::Release()
